streamlit/requirements.txt

numpy==1.19.2
Pillow==7.2.0
streamlit==0.67.1
fastai==2.0.10
fastcore==1.0.9
fastprogress==1.0.0
https://download.pytorch.org/whl/cpu/torch-1.6.0%2Bcpu-cp36-cp36m-linux_x86_64.whl
